+++ docs/changelog.md
@@ -119,3 +119,4 @@
 - Git Push [2025-10-19 16:05:30](.logs/audits/git/git.html#run-20251019-160530) → https://github.com/gunreip/tafeld.git [main] : ProjectArtisanDocs integration
 - Git Push [2025-10-19 20:41:47](.logs/audits/git/git.html#run-20251019-204147) → https://github.com/gunreip/tafeld.git [main] : Automated commit
 - Git Push [2025-10-19 21:18:41](.logs/audits/git/git.html#run-20251019-211841) → https://github.com/gunreip/tafeld.git [main] : Automated commit
+- Git Push [2025-10-23 16:52:17](.logs/audits/git/git.html#run-20251023-165217) → https://github.com/gunreip/tafeld.git [main] : Automated commit